backend/requirements.txt
# backend/requirements.txt

Flask>=2.2.0
Flask-Cors>=3.0.10
Flask-SQLAlchemy>=2.5.1
Flask-JWT-Extended>=4.4.4
gunicorn>=20.1.0
qrcode[pil]>=7.4
python-dotenv>=0.21.0
psycopg2-binary>=2.9.6
